Move SimpleCache class to its own module
